取消Url类 并入Route类的buildUrl方法
